Flytt ProsesseringFlyt til egen fil
